requirements.txt
shapely == 1.8.4
uvicorn == 0.20.0
fastapi == 0.92.0
requests == 2.28.1
sqlalchemy == 1.4.45
geoalchemy2 == 0.12.5
psycopg2 == 2.9.3; sys_platform == 'win32'
psycopg2-binary; sys_platform != 'win32'
python-dotenv == 0.21.0
